ToDotNot unit test conversion - part 7.
