+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesDateTimeValuedProperties/Schema.json
@@ -1,0 +1,9 @@
+{
+  "type": "object",
+  "properties": {
+  "startTime": {
+    "type": "string",
+    "format": "date-time"
+    }
+  }
+}
