+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesUriValuedPropertiesFromUriReferenceFormat/ExpectedClass.cs
@@ -1,0 +1,20 @@
+using System;
+using System.CodeDom.Compiler;
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+
+namespace N
+{
+    [DataContract]
+    [GeneratedCode("Microsoft.Json.Schema.ToDotNet", "$JSchemaFileVersion$")]
+    public partial class C
+    {
+        public static IEqualityComparer<C> ValueComparer => CEqualityComparer.Instance;
+
+        public bool ValueEquals(C other) => ValueComparer.Equals(this, other);
+        public int ValueGetHashCode() => ValueComparer.GetHashCode(this);
+
+        [DataMember(Name = "targetFile", IsRequired = false, EmitDefaultValue = false)]
+        public Uri TargetFile { get; set; }
+    }
+}
