+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesUriValuedPropertiesFromUriReferenceFormat/Schema.json
@@ -1,0 +1,9 @@
+{
+  "type": "object",
+  "properties": {
+  "targetFile": {
+    "type": "string",
+    "format": "uri-reference"
+    }
+  }
+}
